requirements.txt
# This is not a good requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
torch
pandas
mlflow
info-nce-pytorch
# For attention visualization
seaborn
